// packages/mods/src/schema.ts
export type Tag = string;

export interface ModTypeProps {
  primary: number;
}

export interface SpawnWeightProps {
  value: number;
  tag: Tag;
}

export interface StatProps {
  id: string;
}

export interface ModProps {
  id: string;
  level: number;
  domain: number;
  name: string;
  correct_group: string;
  generation_type: number;
  stat1_min: number;
  stat1_max: number;
  stat2_min: number;
  stat2_max: number;
  stat3_min: number;
  stat3_max: number;
  stat4_min: number;
  stat4_max: number;
  stat5_min: number;
  stat5_max: number;
  mod_type: ModTypeProps;
  spawn_weights: SpawnWeightProps[];
  stats: StatProps[];
  tags: Tag[];
}

export interface WeaponTypeProps {
  critical: number;
  speed: number;
  // phys
  damage_min: number;
  damage_max: number;
  range_max: number;
}

export interface AttributeRequirementProps {
  req_str: number;
  req_dex: number;
  req_int: number;
}

export interface ArmourProps {
  armour: number;
  evasion: number;
  energy_shield: number;
}

export interface ShieldTypeProps {
  block: number;
}
// ...
export interface BaseItemTypeProps {
  id: string;
  name: string;
  width: number;
  height: number;
  drop_level: number;
  inherits_from: string;
  weapon_type?: WeaponTypeProps;
  component_attribute_requirement?: AttributeRequirementProps;
  component_armour?: ArmourProps;
  shield_type?: ShieldTypeProps;
  implicit_mods: ModProps[];
  item_class: string;
  tags: Tag[];
}
// ...
export interface EssenceProps {
  primary: number;
  level: number;
  base_item_type: BaseItemTypeProps;
  essence_type: {
    id: string;
    essence_type: number;
    is_corrupted_essence: boolean;
  };
  item_level_restriction: number;
  quiver_mod?: ModProps;
  amulet_mod?: ModProps;
  ring_mod?: ModProps;
  belt_mod?: ModProps;
  shield_mod?: ModProps;
  helmet_mod?: ModProps;
  body_armour_mod?: ModProps;
  boots_mod?: ModProps;
  gloves_mod?: ModProps;
  bow_mod?: ModProps;
  wand_mod?: ModProps;
  staff_mod?: ModProps;
  two_hand_sword_mod?: ModProps;
  two_hand_axe_mod?: ModProps;
  two_hand_mace_mod?: ModProps;
  claw_mod?: ModProps;
  dagger_mod?: ModProps;
  one_hand_sword_mod?: ModProps;
  one_hand_thrusting_sword_mod?: ModProps;
  one_hand_axe_mod?: ModProps;
  one_hand_mace_mod?: ModProps;
  sceptre_mod?: ModProps;
}